Update Laboratory2.Rmd
Analysis. 
We randomize the whole dataset before splitting it into train and test data.  It makes no sense to randomize it after splitting the data.
In part 2, we chose to remove one highly correlated variable, and the accuracy was slightly improved compared to the first model.
In the last part, when we implement the method on News Popularity data, it gives a perfect prediction of 100% accuracy, which is much better than the result we got in lab 1.